Initial commit
A few basic things have been added to kick things off:
-- A screen that adds the game panel (the main component of the game). I
might add a game class that does all the logic.
-- A game panel that draws a grid and colors the cells as needed
...and a few other basic classes for abstraction and manipulation

The goal is to not only create Snake for a person to play, but also to
make an AI that "evolves" to get better at the game every round and
hopefully excel at it.