Add AI implementation
My oh my. This is a big commit. The following was added/changed:

-- A snakeAI class that is basically the snake's "brain". It produces
the inputs that are sent in the neural network. It then controls the
snake based on the predicted action from the network

-- A gameAI class that controls all the snakes in a population, updates
them and their games. I'll add a Generation class to more easily control
that stuff (and add mutation, etc.)

-- A button to speed up the game

-- The neural network was changed to just use one hidden layer and one
output layer, each with their own number of nodes

Note: This AI "works", but there is no mutation implemented yet, so it
will generate new weights (genetics) for each snake every generation