Add option to load an evolution from file or start a new evolution
This happens at the start of the game if the user wants to start an
evolution. If the user picks to load from the file, the data will be
loaded into the first generation (if available). If the user picks to
start a new generation, the existing file will be deleted and
overwritten by the new generation.